requirements.txt
# Telegram Bot Framework
python-telegram-bot[job-queue]==21.10

# Groq AI Integration
groq==1.1.2

# Scheduling
APScheduler==3.10.4

# Environment Variables
python-dotenv==1.0.0

# Timezone support
pytz==2024.1

# Turso / libSQL (cloud database via HTTP API)
requests==2.31.0

# Health server for Render
aiohttp==3.9.3
